feat(demixer): handle DemixerApi returning errors as JSON
